backend/requirements.txt.txt
Flask
pymongo[srv]
python-dotenv
Flask-Cors
google-auth
PyJWT